+++ wflow-py/wflow/wflow_routing.py
@@ -242,8 +242,8 @@ class WflowModel(DynamicModel):
         # Set and get defaults from ConfigFile here ###################################
         self.maxitsupply = int(configget(self.config, "model", "maxitsupply", "5"))
         # max number of iteration in abstraction calculations
-        self.reinit = int(configget(self.config, "model", "reinit", "0"))
-        self.fewsrun = int(configget(self.config, "model", "fewsrun", "0"))
+        self.reinit = int(configget(self.config, "run", "reinit", "0"))
+        self.fewsrun = int(configget(self.config, "run", "fewsrun", "0"))
         self.OverWriteInit = int(configget(self.config, "model", "OverWriteInit", "0"))
         self.updating = int(configget(self.config, "model", "updating", "0"))
         self.updateFile = configget(self.config, "model", "updateFile", "no_set")